Release 5.3.1897
Updated the definition of swath format MBF_LDEOIH (format 71),
which is used for the "fast bathymetry" or *.fbt files in MB-System
processing environments. The previous form of *.fbt files had a
serious limitation in that bathymetry from multibeams operated
near the seafloor in deep water did not represent the full
numerical resolution of those data. The updated format allows for
depth and distance resolution to 0.001 m even in the deep ocean.
Old *.fbt files are read transparently, but newly written files
will be in the new form.
IMPORTANT: In order to avoid compatibility issues, all users 
operating on the same projects should upgrade to this version or
higher once one user upgrades.

Updated support for Reson 7k data to include the "1022" record.
We still don't know what this is, but the code now reads and 
writes it without complaint.