Provided a means to continue using the old definition of swath format MBF_LDEOIH (format 71) for fbt files. Users can now use mbdefaults to set the format 71 version that will be generated when fbt files are created by mbdatalist and mbprocess. If  	mbdefaults -Fold is executed, the .mbio_defaults file in the user's home directory will be altered so that all new fbt files are in the old format. The purpose of this backwards compatibility is to allow users to take advantage of the Fledermaus capability for editing bathymetry in MB-System processing environments. This involved changes to mbcopy as well as to the MBF_LDEOIH i/o module.
Changed the swath data import code so that sections will be created even
when all beams have been flagged as bad. This is to allow surveys
to remain intact, with no breaks, because the concept of a continuous
survey is increasingly important in MBnavadjust. Also altered the 
status displays and listings to show survey id along with file and
section id's.