Mbnavadjustmerge: Added ability to unset (delete) global ties and crossing ties. The relevant mbnavadjustmerge commands are:     --unset-global-tie=file:section     --unset-tie=file1:section1/file2:section2
Mbnavadjust, mbnavadjustmerge: Explicitly added references to libgmt and libpsl in
the src/mbnavadjust/Makefile.am file so that the mbnavadjust and mbnavadjustmerge
executables are linked to those libraries. This is to solve a mysterious dependency
of mbnavadjust on libgmt on CentOs6 - the program was failing due inability to find
libgmt.so.5 even though the code does not reference any GMT or GMT5 function or header
file. Presumably this change will go away when the mystery is solved.