Mbedit, mbnavedit, mbvelocitytool, mbgrdviz, mbeditviz: Removed call to X11 function XtSetLanguageProc() in all graphical tools. This call apparently causes the program locale to be reset according system defaults, even if the program implicitly begins in the standard C locale. If the system locale uses "," rather than "." as the decimal point marker (e.g. a normal European locale), then a GMT 5 initialization will probably fail on reading fonts, at least on many if not all Linux distributions. Thanks to Giancarlo Troni for figuring this one out.
General: Added mbio/mb_platform.c to the archive. This source file includes the
functions to read, write, and use platform, sensor, and sensor offset definitions.
At this time the mb_platform_*() functions are only used in mb7kpreprocess, but
they will soon be used in mbprocess and all the preprocess programs.

Mbprocess: changed application of tide corrections so that these corrections are
applied to the platform depth values rather than the bathymetry values. The result
is the same, but now the navigation (or trajectory) of the processed files is
corrected in addition to the bathymetry.