Format 201 (MBF_HYSWEEP1): Changed handling of file header data to enable the extraction of platform definition data by the program mbmakeplatform.
MBmakeplatform: Now capable of extracting the list of sensors and the associated
offsets from Hypack HSX files, and outputting an MB-System platform file.

MBhysweeppreprocess: Now allows for sensor offsets to be defined by an
MB-System platform file, which can be extracted from HSX data files with
mbmakeplatform, and can be modified and augmented using the same program.

MBeditviz: Added a new sounding filter mode that calculates the sounding density
for a 3D "voxel" representation of the selected volume, and then flags soundings
in voxels containing less than a threshold of soundings. This algorithm flags
isolated soundings, an approach that is particularly useful for submarine
lidar data. This filter is accessible from the "action" menu of the 3D sounding
window.