Formats 58 and 59 (MBF_EM710RAW and MBF_EM710MBA): Fixed handling of sensordepth and heave data during preprocessing. Program mbkongsbergpreprocess is now deprecated and will be removed from the distribution soon. Users should now use mbpreprocess for preprocessing Kongsberg multibeam data.
Format 88 (MBF_RESON7KR): Fixed bug in handling of TVG data records.

Mbdatalist: Added a new option -D that causes mbdatalist to list the unique
datalist files referenced as a recursive datalist structure is parsed. The
recursion level of each datalist file is indicated in the listing.

Mblist and mbnavlist: Augmented the -Zsegmenttag option for printing out segment
start lines in the output. Normally the string segmenttag is a single character
that is recognized by GMT as a segment beginning when parsing tabular data. Now,
if segmenttag == "swathfile" the segment start lines will consist of the character
'#' followed by the path to the source swath file, and if segmenttag == "datalist"
the segment start lines will consist of the character '#' followed by the path
to the source datalist file