MBpreprocess: Fixed generation of *.baa, *.bsa, and *.bah files so they include only data relevant to the associated swath file, which means samples from 10 seconds before the first survey data through 10 seconds after the last survey data. The program now removes any pre-existing *.ata, *.sta, *.ath, *.baa, *.bsa, and *.bah files.
MBconfig: New program to provide command line access to the MB-System version
and the locations of the Levitus database and the OTPS tide prediction software.

MBlevitus: The variable storing the location of the Levitus database file has
been changed from "ifile" to "levitusfile".

Datalist parsing: The mb_datalist_read() function changed with the 5.5.2299
revision, including adding a new argument. The old version of that function is
now available within the API as mb_datalist_readorg().