General: Fixed problem with data record type informational messages
General: Fixed problem with time gap nonfatal error detection

MBpreprocess: Added --skip-existing option that causes MBpreprocess to skip
preprocessing of files referenced in the input datalist that have already been
preprocessed. This allows efficiently adding new files to a survey, particularly
in a shipboard context where one is trying to process new data as they are
collected.

MBotps: Added -S option that causes MBotps to skip calculating a tidal correction
model for files referenced in the input datalist that already have models calculated
by mbotps. The skipped files will not be viewed as out-of-date by mbprocess, and
thus will not be reprocessed if mbprocess or mbm_multiprocess are run.
This allows efficiently adding new files to a survey, particularly
in a shipboard context where one is trying to process new data as they are
collected.

MBgrdviz: Added capability to output site data as Winfrog waypoint (*.WPT) files.

MBeditviz: Added capability to optimize an additional bias parameter (sort of...)
in the 3D sounding view. The new parameter is a "snell correction", referring to
adjusting the receive beam angles of multibeam data by applying a Snell's law
correction equivalent to changing the speed of sound used for beam forming. This
parameter is expressed as a ratio of the new to original sound speed, so that
a snell correction value of 1.00 corresponds to no change, a value of 0.995
corresponds to a slight reduction in the sound speed, and a value 1.005
corresponds to a slight increase in the sound speed. This correction is intended
for near-bottom surveys from submerged platforms (ROVs or AUVs) where sound speed
is calculated from CTD data (and values may plausibly be slightly erroneous) and
where errors due to raytracing through an incorrect model of the water column
sound speed are unlikely to be significant. This correction is generally not useful
for multibeam data collected from a surface vessel.

MBminirovnav: Added new program that processes raw MBARI MiniROV data logs to
produce standard MBARI ROV navigation files (format MBF_MBARIROV)
MBpreprocess: Add option to skip preprocessing files in the input datalist that
have already been fully preprocessed.

MBsegygrid: Added a trace mode=2 to the -Smode[/start/end[/schan/echan]] command
for which start and end refer to shot numbers and all traces within that range of
shot numbers are plotted. Effectively  the channel numbers are ignored by setting
all trace channels to 0 and plotting channel 0.

MBsslayout: Changed arguments to have '-' rather than '_' characters between
words in multiword commands (e.g. --topo_grid_file is now --topo-grid-file).
Also now have code identifying the fileroot of the input file and using that in
the construction of output file names so that the input format suffix is not
embedded in the middle of the output filename. Also improved the swathplotting
script generated by mbsslayout.

Formats 132 (MBF_EDGJSTAR) and 133 (MBF_EDGJSR2): Fixed preprocessing function in
the i/o module so that altitude values read in by mbpreprocess are actually applied
in the output file.

Format 160 (MBF_SEGYSEGY): Added recognition of this format by file suffixes
*.SEG and *.seg (in addition to *.SEGY and *.segy).

MBm_makedatalist and Formats 56, 57, 58 59 (MBF_EM300RAW, MBF_EM300MBA, MBF_EM710RAW,
MBF_EM710MBA): The mbm_makedatalist macro will now attempt to sort Kongsberg data
into time order based on filenames when the file suffixes follow the MB-System
convention (*.mb56, *.mb57, *.mb58, *.mb59). Previously this functionality was
only applied to files with the *.all suffix. This capability is applied by default
but can be disabled with the -T option.