Update 5.6.20181016
-----> 5.6.20181016 (October 16, 2018)

Many files: Changed type of pingnumber variable from int to unsigned int.

MBprocess: Fixed problem with the creation of *.resf files in which the edit
events did not properly include the multiplicity factor required for data
from multi-head sensors.

MBvoxelclean: New program mbvoxelclean identifies and flags artifacts in swath
bathymetry data using a 3D density filter. The notion applied is that true targets
(e.g. the seafloor) result in dense regions of soundings while sparse soundings
in the water column or the subsurface are erroneous and can be flagged as bad.
This technique is more appropriate for lidar data than multibeam sonar data.
The resulting sounding edit events are output to edit save files which can be
applied to the data by the program mbprocess. These are the same edit save
files created and/or modified by mbedit, mbeditviz, mbedit,
and \fBmbclean\fP.