5.7.5beta2 (March 4, 2019)
Format 261 (MBF_KEMKMALL): There is progress towards support for the new
Kongsberg kmall format. Only EM2040 data supplied by Kongsberg is being
used for testing with mbinfo, mbcopy, and mbprocess, so far. The memory
management faults are fixed, but it isn't all working yet.

Mbcopy: Changed logic so that in a full copy all data records are passed on to the
output even if they generate nonfatal errors (e.g. out of time or location bounds).