5.7.5beta6
-----> 5.7.5beta6 (March 26, 2019)

Build system: Modified configure.ac to fix a problem building with the new
Proj 6.0.0. MB-System uses a longstanding Proj API that is deprecated in Proj 6.
The relevant header file is proj_api.h, which can only be used if compiled
with the preprocessor macro ACCEPT_USE_OF_DEPRECATED_PROJ_API_H set. This compiler
flag has been added to the autoconf test for usability of this header file.