Introduce testing of command line apps using python3 starting with partial coverage of mbinfo.
I picked python 3.6 as the minimum as it's been out for a while and
py2 is End Of Life.

This captures some quirks in the json output with leading spaces that
should eventually be removed.

I used a hydrosweep file from my 2001 cruise and cut it down by hand
to be reasonably small.  The original is here:

http://vislab-ccom.unh.edu/~schwehr/TTN136B/Data/hysweep/