Make all headers have include guards (mbio only).
Use the pattern FILENAME_H_ for the define.  I skipped using a
filepath or package specific prefix as the files are pretty unique
with "MB" in front. The one exception what projection.h, which is too
generic a name to be safe to use.  This makes it easier to grep the
guards.

Removes some guards wrapped around `#include` lines that violate the
concept of Don't Repeat Yourself (DRY).

Fixed mbsys_reson7k3.h that had missing quotes on the include:

    #include stdint.h

See also:
- https://google.github.io/styleguide/cppguide.html#The__define_Guard
- https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#Rs-guards