Add mbnavlist_test.py.
Using an exact match on the default output of mbnavlist is pretty brittle,
but it does make it easy to get a format tested.

Note that the generation of text for error messages is different
between mbinfo and mbnavlist.  mbinfo by default sends things to
stdout whereas mbnavlist always uses stderr.