utilities: Fix up includes in iwyu (include what you use) order. (#120)
Sort the includes alphabetically in two groups:
1. C std/system includes
2. All other includes

Also:

Remove redundant comments before includes and #defines.

Remove location comments.  It is fun to see where the code was written
on the Earth, but that information isn't useful for a person working
on the code.  My take is that kind of info should go in a commit
message if it provides context to a change.