Add mbinfo_test.py test case for mb11.
This is my first attempt at using mbcopy to create super simple test
cases for other sonar formats.  This is not as good as using data
written by the native sonar software, but it is a start.  The
TN136HS.309.snipped file is about as simple as can be, so it makes sure
that there is a very basic test case available.

Generated by doing:

  mbcopy -F21/11 -Itest/utilities/testdata/mb21/TN136HS.309.snipped.mb21 -Otest/utilities/testdata/mb11/TN136HS.309.snipped.mb11
  mbinfo -X1 -ITN136HS.309.snipped.mb11 > TN136HS.309.snipped.mb11.json