5.7.6beta3
Integration with PROJ: The API for the PROJ coordinate transformation library
libproj has changed with the release of version 6. The build system now tests
which version of PROJ is installed and builds MB-System using the available
API.

Mbgrid: Mbgrid had an undocumented ability to set the interpolation tension using
a third argument to the -C option. This has been removed. The tension value is set
with the -T option, consistent with the documentation.

Mbgrid and mbmosaic: The functions used by mbgrid and mbmosaic to integrate with
GMT's i/o capabilities to write the output grids have been augmented to embed
the geographic coordinates in the fashion supported by GDAL and many software
packages that utilize it.

Mbcontour, mbswath, mbgrdtiff: The preprocessor commands in the source files
for the GMT modules in MB-System have been updated to handle the existence
of versions 5.0 through 6.0+.

Code stye: Kurt Schwehr is systematically altering the code to conform to best
practices regarding header inclusion and reduced variable scope, particularly
for loop indices.