For beta release 5.7.6beta3
Format MBF_GSFGENMB (format 121, Generic Sensor Format, GSF): The version of
libgsf used to read and write Generic Sensor Format (GSF) files has been
updated to version 3.0.9. The libgsf library code is released by Leidos under
a LGPL2.1 license, and is included directly into the MB-System source code with
minor changes. For 3.0.9 the only changes required are to avoid redefinition
of several key system functions (e.g. fseek); these redefinitions are avoided
using a preprocessor command that is only activated in the MB-System build
system. The command
  mbformat -F121
now outputs the version of the GSF library embedded in MB-System as a part of
the format description.