Setting up for release 5.7.6.beta8
There was a problem with 5.7.6beta7 building due to a missing #include in one source file. I deleted that release, updated the code, and remade the release with the same name. It turned out that the description was updated, but Github did not update the broken tarball. Therefore, I'm incrementing the release and it's tag.