5.7.6beta16
-----> 5.7.6beta16 (October 29, 2019)

MBnavadjust: Fixed bug in calculating the range of contour values and the size
of the memory allocation for contours.

Code stye: Kurt Schwehr is systematically altering the code to conform to best
practices regarding header inclusion and reduced variable scope, particularly
for loop indices. Current changes include replacing MB_YES/MB_NO with boolean
true and false, and changing the type of the associated variables from int to
bool.