Undoing prior commit - now just not building mbsvpselect without PROJ6
Found that prior commit still did not build on CentOs 7 with proj 4.9
Giving up on supporting geodesic functions when built with PROJ4 - now in this case mbsvpselect will not get built.
This is because I do not want to return to the prior situation where old copies of geodetic.h and geodetic.c taken from the Proj distribution were included in the utilities directory. This functionality is now part of PROJ6, and the header file and the functions should be obtained from the Proj installation.