Change to handling Proj versions
Going back to approach in which MB-System starts using the Proj 6 API when the function proj_normalize_for_visualization() is available, which started with Proj release 6.1.0. If proj_normalize_for_visualization() is not available, MB-System will use the deprecated Proj 4 API. This effectively bypasses the API changes made for Proj releases 5.0 through 6.0.