Prep for 5.7.6.beta32
Mbeditviz: Fixed problem with preserving changes to edits that are recorded in
*.esf files.

Code style: Kurt Schwehr is systematically altering the code to conform to best
practices and adding build tests. The improvements included in this beta release
include work on mbview and programs using mbview.

Code style: Tom O'Reilly and David Caress added README.md files in each of the subdirectories under src/.