Support old/new versions of mbgrdtiff related to upcoming GMT 6.1
Paul Wessel contributed a new version of mbgrdtiff that constructs the image by a direct call to the grdimage module. This version will only build and work with GMT 6.1 and later. The original code included code building the image, duplicating functionality in grdimage. The new version is now mbsystem/src/gmt/mbgrdtiff.c and the original version mbsystem/src/gmt/mbgrdtifforg.c. The build system has been augmented to check the GMT version and build mbgrdtiff from the original when that is <6.1 and the new file when that is >= 6.1.