Fix mbgrid and add Qt5 support to build system
Mbgrid: Fixed failure of the two gridding algorithm using beam footprints in
which the footprint size blows up at low grazing angles. The fix is to not
use soundings for which the grazing angle is less than five degrees.

Mbgrid and mbmosaic: Replaced int values with bool values where appropriate.

Build system: Added option to enable building tools based on Qt5 to the autotools
build system. If the configure script is given the --enable-qt option, it
checks if Qt5 has been installed, and if so, determines the compiler and linker
flags required to build tools using Qt5. At present, the only Qt5 based code
is in the directory src/qttest, and this code does not successfully compile
and run. Therefore it is not recommended to run configure with the --enable-qt
option. However, the build system is now ready to support Qt5 based tools as
they are developed.