mbedit - address issue #1041
Addressed TODO's left by Kurt in mbedit_prog.c in conjunction with issue #1041. The always true else statements have been commented out. Some variables converted to bool as suggested.  The logic questioned for the reverse cut filters is correct and has been left in place. When begin < end then the cut is in the interior of the swath. When begin > end then cuts are made at the outer edges of the swath.