Build system work has Qt5 test (almost) working on Mac
Many changes to the build system to get Tom O'Reilly's Qt5 demo to work (almost). The challenge is dealing with the Qt frameworks and C++ code generation at build time.
1. There is a new m4 macro in mbsystem/m4/ax_have_qt_mb.m4 which was obtained from Gnu (originally ax_have_qt.m4 circa 2014) and modified to work with MB-System and the current Qt5. This is called from the configure.ac file and both detects a working Qt installation and obtains the compile and link flags necessary to build Qt code.
2. A lot of the Qt UI code is generated by Qt at build time through the utility qmake and tools it calls like moc, uic, and rcc. However, we want the actual compiling and linking to be handled by make using Makefiles generated by Autotools. The draft solution adopted here is to call qmake from the configure script such that it generates a Makefile named Makefile.qmake in the relevant source directory. We then craft the Makefile.am used by Autotools so that the Makefile referenced by running make all can generate the needed source files through calls like:
	make -f Makefile.qmake compiler_rcc_make_all
	make -f Makefile.qmake compiler_moc_header_make_all
that just call the code generation tools in the Qt package and don't actually compile anything. Once the source files have been generated, they are then compiled and linked normally.
At this point there is one Qt test program named mbqttest that is build in the source directory src/qttest/ if and only if configure is called with the undocumented option --enable-qt.
On a Mac this almost works. The program mbqttest is in fact built in the source directory, and that binary in place works. However, when the binary is installed with make install, the installed program fails because it calls libraries that are not accessible through the embedded rpath. This is the next problem to solve.
An mbqttest binary built solely using the qmake Makefile is in fact installable, so the key if to figure out what is different for the rpath between the two cases.