Progress with build system for Qt5 based tools
Build system for Qt5 based tools: Improved the build system handling of Qt5
programs. The test program enabled by running configure with the --enable-qt
option has been renamed mbgrdvizqt, but still resides in src/qttest. The
approach now is to do the following:
(1) The configure script creates the Qt project file src/qttest/mbgrdvizqt.pro
    and then modifies the values of variables specifying the required MB-System
    and GMT libraries and headers.
(2) The configure script runs qmake in src/qttest to generate a Makefile named
    src/qttest/Makefile.qmake
(3) The configure script generates a Makefile in src/qttest that in turn will
    run make itself using Makefile.qmake for targets all, install, uninstall,
    and clean.
This approach now works on a Mac. It's time to test on Linux.