Fix to build with GMT 6.1.0 & tweak to mbnavadjust
Mbgrd3obj: Had to add #ifdef's on GMT version because the number of parameters
passed to gmt_init_module() changes between GMT 6.0 and 6.1.

Mbnavadjust: Changed calculation of overlap for a crossing so that the value is
not small in the case that a small section (by area) is entirely inside the bounds of a
much larger section. This case happens when adjusting navigation from surveys at
at two very different scales (e.g. AUV survey with ship hull mounted survey,
or ROV survey with AUV survey).