Temporary debugging tweak to CI tests
Changed test/utilities/Makefile.in so that the CI test log will cat the log files for the tests of mbbackangle, mbgrid, and mbprocess. These are failing in one of the Travis-CI test builds - the one for Ubuntu Xenial. All other test builds succeed, and the reason for the one failure is unknown for lack of access to the failed test output.