mbnavadjust inversion algorithm fix
Mbnavadjust: Fixed problems in the inversion algorithm.
The complexity of this algorithm results from the tendency
of overdetermined least squares solutions to work best at providing uniform scale,
zero mean solutions. Consequently we construct the solution in three steps.
First we solve for the average offsets
between continuous surveys, and make this the starting model for the second step.
Then we solve for a coarse ("chunk") perturbation model in which each chunk is a
dataset consisting of about five sections worth of continuous survey data. This is
done through an iterative damped relaxation procedure, and the resulting perturbation
model is added to the starting model for the third and final step. The final step
is to solve for a perturbation satisfying all of the navigation ties at full
resolution. The problem is set up as an overdetermined least squares matrix
problem and is solved using LSQR. The regularization is just penalizing the
first derivative of the perturbation at present - we are not currently penalizing
the second derivative.

Mbnavadjust: Fixed crash on Linux when executing the Auto Set Vertical Offset
function (didn't dimension arrays large enough).