Work on Kongsberg multibeam formats
Format MBF_KEMKMALL (261): Brought reading and writing of *.kmall files up to
revision G of the Kongsberg data format specification. Also fixed problems in the
logic relating calculated depths as stored in the format to the depth values
reported by the MB-System API functions, and with the beam angles calculated for
use in recalculating bathymetry through raytracing. However, problems remain
with extracting the installation bias parameters and time latencies that must
be applied to the asynchronous ancillary data in order to properly calculate
these values. At present, recalculation of bathymetry by raytracing still does
not work correctly with this format - in some cases additional attitude biases
and time latencies must be applied, and in others the result is simply not right.

Formats MBF_EM710RAW (58) and MBF_EM710MBA (59): Added allowed sonar models
- unexpectedly we are seeing *.all files from some current generation sonar
installations, so we have to allow EM304 and EM712 sonar models with the
obsolete format.