Changes to build with GMT 6.1.0 and GMT 6.2
The GMT 6.1.0 release contains a change to the API that broke the build of the GMT module mbswath. It also contained changes to the module initialization code that caused mbswath, mbcontour, and mbgrdtiff to crash. The code has been modified with (more) proprocessor directives determined by the installed GMT version to build with GMT versions 5.2, 5.4, 6.0, 6.1, and the current master that will become 6.2. For GMT versions >= 6.0 the MB-System code no longer depends on either the gmt/src/gmt_mb.h header file or the obsolete gmt_begin_module() function.

Also corrected the usage statement for mbphotomosaic, mbgetphotocorrection, and mbphotogrammetry.