Change configure.ac to fix problem with Homebrew installation
I added the line
  prefix=$ac_default_prefix
near the top of configure.ac in order to insure the installation prefix is embedded into mbio/mb_config.h to be accessed by mbconfig. This appears to have the effect of overriding the default prefix for Homebrew installations, which makes the install phase fail because Homebrew installations go into /usr/local/Cellar/package/ not directly into /usr/local.