Fixes to mblist and mbphotomosaic
Mblist: Fixed bug in which output in the CDL format was broken, which in turn broke output in netCDF.

Mbphotomosaic: Fixed generation of photomosaics in projected coordinate systems (e.g. UTM).

Mbphotomosaic, mbgetphotocorrection, mbphotogrammetry: Changed source file names from the *.cpp suffix to the *.cc suffix to be consistent with other C++ source files in MB-System.