Mods preparing for 5.7.7beta07
Mbclean: Fixed two bugs. The first involved setting the left/right bounds of
flagging by acrosstrack distance using the -Y option, and the other resulted in
slope flagging being applied when not requested when the -Y option was used alone.

Mbpreprocess: Added option to set directory where output files are written instead
of writing the output files parallel to the input swath files.

Format 121 (MBF_GSFGENMB): Modified handling of multibeam snippet backscatter so
that samples associated with beams flagged as bad are nulled, and so that modified
backscatter is written to the output file by mbprocess.

Mbeditviz: Modified 3D sounding view so that the depth buffering is actually used.

Mbm_makedatalist: Now ignores the most recent file when the -L option is used.