Fixes to mbgrid, mbmosaic, mbnavadjust, mbgrdviz, mbm_grdplot
Mbgrid and mbmosaic: Fixed setting output grid format with the -G option. The
definition and parsing of the options is now consistent with GMT 6.

Mbnavadjust: Added option to unset skipped crossings.

Mbnavadjust: Changed vertical scaling of the time series model plot and the
tie plots so that these are centered on the midpoint values rather than zero.

Mbgrdviz: Changed how routes are output as Risi survey scripts - now the lateral
positions of waypoints are relative east and north meters in UTM rather than
local x,y calculated using mtodeglon and mtodeglat from a simple spheroid.

Mbm_grdplot: Fixed so that it does not crash if the input grid has either no
valid values (all are NaN) or all valid values are the same.